Fix : 달력 반복 조회 관련 수정(S10P31A602-572)

See merge request s10-final/S10P31A602!196

+++ backend/src/main/java/com/butter/wypl/schedule/data/request/RepetitionRequest.java
@@ -17,7 +17,7 @@ public record RepetitionRequest(
 	LocalDate repetitionEndDate,
 
 	@JsonProperty("day_of_week")
-	byte dayOfWeek,
+	int dayOfWeek,
 	Integer week
 
 ) {
